finalizada chamada da api de searchUsers
