+++ src/services/Linkr.js
@@ -58,8 +58,8 @@ function sendEditPostRequest (postId, text, token){
     return axios.put(`${BASE_URL}/posts/${postId}`, {text: text}, createConfig(token));
 }
 
-function searchUser (queryStrConfig) {
-    return axios.get(`${BASE_URL}/users/search`,queryStrConfig)
+function searchUser (queryStr, token) {
+    return axios.get(`${BASE_URL}/users/search?username=${queryStr}`,createConfig(token))
 }
 
 export {
